refactor: improve DirectClient API routing and error handling
- Separate API route handlers into dedicated `handlers.ts` file
- Add custom error classes in `errors.ts` for better error management
- Refactor message, whisper, speak, and image route handlers with improved error handling
- Implement centralized runtime retrieval method in DirectClient
- Enhance response generation and message processing logic
