+++ packages/client-direct/src/errors.ts
@@ -1,0 +1,11 @@
+export class AgentNotFound extends Error {
+    constructor(message: string = "Agent not found") {
+        super(message);
+    }
+}
+
+export class NoResponseError extends Error {
+    constructor(message: string = "No response from generateMessageResponse") {
+        super(message);
+    }
+}
