+++ packages/client-direct/src/index.ts
@@ -1,124 +1,1 @@
-import bodyParser from "body-parser";
-import cors from "cors";
-import express from "express";
-import multer from "multer";
-import {
-    elizaLogger,
-    AgentRuntime,
-    Client,
-    IAgentRuntime,
-    settings,
-} from "@elizaos/core";
-import * as fs from "fs";
-import * as path from "path";
-
-import { createApiRouter } from "./api.ts";
-
-const storage = multer.diskStorage({
-    destination: (req, file, cb) => {
-        const uploadDir = path.join(process.cwd(), "data", "uploads");
-        // Create the directory if it doesn't exist
-        if (!fs.existsSync(uploadDir)) {
-            fs.mkdirSync(uploadDir, { recursive: true });
-        }
-        cb(null, uploadDir);
-    },
-    filename: (req, file, cb) => {
-        const uniqueSuffix = `${Date.now()}-${Math.round(Math.random() * 1e9)}`;
-        cb(null, `${uniqueSuffix}-${file.originalname}`);
-    },
-});
-
-export class DirectClient {
-    public app: express.Application;
-    private agents: Map<string, AgentRuntime>; // container management
-    private server: any; // Store server instance
-    public startAgent: Function; // Store startAgent functor
-    public upload = multer({ storage });
-
-    constructor() {
-        elizaLogger.log("DirectClient constructor");
-        this.app = express();
-        this.app.use(cors());
-        this.agents = new Map();
-
-        this.app.use(bodyParser.json());
-        this.app.use(bodyParser.urlencoded({ extended: true }));
-
-        // Serve both uploads and generated images
-        this.app.use(
-            "/media/uploads",
-            express.static(path.join(process.cwd(), "/data/uploads"))
-        );
-        this.app.use(
-            "/media/generated",
-            express.static(path.join(process.cwd(), "/generatedImages"))
-        );
-
-        const apiRouter = createApiRouter(this.agents, this);
-        this.app.use(apiRouter);
-    }
-
-    // agent/src/index.ts:startAgent calls this
-    public registerAgent(runtime: AgentRuntime) {
-        this.agents.set(runtime.agentId, runtime);
-    }
-
-    public unregisterAgent(runtime: AgentRuntime) {
-        this.agents.delete(runtime.agentId);
-    }
-
-    public start(port: number) {
-        this.server = this.app.listen(port, () => {
-            elizaLogger.success(
-                `REST API bound to 0.0.0.0:${port}. If running locally, access it at http://localhost:${port}.`
-            );
-        });
-
-        // Handle graceful shutdown
-        const gracefulShutdown = () => {
-            elizaLogger.log("Received shutdown signal, closing server...");
-            this.server.close(() => {
-                elizaLogger.success("Server closed successfully");
-                process.exit(0);
-            });
-
-            // Force close after 5 seconds if server hasn't closed
-            setTimeout(() => {
-                elizaLogger.error(
-                    "Could not close connections in time, forcefully shutting down"
-                );
-                process.exit(1);
-            }, 5000);
-        };
-
-        // Handle different shutdown signals
-        process.on("SIGTERM", gracefulShutdown);
-        process.on("SIGINT", gracefulShutdown);
-    }
-
-    public stop() {
-        if (this.server) {
-            this.server.close(() => {
-                elizaLogger.success("Server stopped");
-            });
-        }
-    }
-}
-
-export const DirectClientInterface: Client = {
-    start: async (_runtime: IAgentRuntime) => {
-        elizaLogger.log("DirectClientInterface start");
-        const client = new DirectClient();
-        const serverPort = parseInt(settings.SERVER_PORT || "3000");
-        client.start(serverPort);
-        return client;
-    },
-    stop: async (_runtime: IAgentRuntime, client?: Client) => {
-        if (client instanceof DirectClient) {
-            client.stop();
-        }
-    },
-};
-
-export default DirectClientInterface;
+export * from "./client";
